// sourcecode/BTrak.Packages/snova-project-management/PM-module-components/pm-components/src/lib/snovasys-projects/models/selectedGoalActivityModel.ts
export class SelectedGoalActivityModel {
    goalId: string;
    isIncludeUserStoryView: boolean;
    isIncludeLogTime: boolean;
    userId: string;
    sprintId: string;
    projectId: string;
    isFromActivity: boolean;
    pageNumber: number;
    pageSize: number;
    dateFrom: any;
    dateTo: any;
}