// sourcecode/BTrak.Packages/snova-project-management/PM-module-components/pm-components/src/lib/snovasys-projects/models/dashboardFilterModel.ts
export class DashboardFilterModel {
    projectId: string;
    userId: string;
    goalId: string;
    sprintId: string;
    dateFrom: string;
    dateTo: string;
    date: Date;
    entityId: string;
    branchId: string;
    designationId: string;
    roleId: string;
    sprintStartdate: Date;
    sprintEndDate: Date;
    departmentId: string;
    isFinancialYear: string;
    isActiveEmployeesOnly: string;
    monthDate: string;
    yearDate: string;
    isListView: boolean;
    auditId: string;
}